Ignore DevTools information in case it's already installed

+++ src/app/Console/Commands/Install.php
@@ -59,12 +59,14 @@ class Install extends Command
         $this->info(' Backpack installation finished.');
 
         // DevTools
-        $this->box('Did you know about Backpack DevTools?');
-        $this->note('DevTools adds a dead-simple web interface to easily generate Models, Migrations, Seeders, Factories, CRUDs, etc.');
-        $this->note('But it\'s a paid tool. For more info, payment and access, please visit https://backpackforlaravel.com/products/devtools');
+        if (! file_exists('vendor/backpack/devtools')) {
+            $this->box('Did you know about Backpack DevTools?');
+            $this->note('DevTools adds a dead-simple web interface to easily generate Models, Migrations, Seeders, Factories, CRUDs, etc.');
+            $this->note('But it\'s a paid tool. For more info, payment and access, please visit https://backpackforlaravel.com/products/devtools');
 
-        if ($this->confirm('Would you like to install Backpack DevTools?', false)) {
-            $this->call('backpack:require:devtools');
+            if ($this->confirm('Would you like to install Backpack DevTools?', false)) {
+                $this->call('backpack:require:devtools');
+            }
         }
     }
 }
